fix errors introduced
Reran on the original data set and there were a couple errors introduced from refactoring things. Need to add a line to the end of a sequence entry, wrote to the temp file instead of outfile in one instance, added more print statements to indicate to user what was happening, cleaning up how the tophat command is printed, add overwriting of existing files that are created for the run. Tested and works again. :)